requirements.txt
langgraph
langchain-core
chromadb
sentence-transformers
openai
python-dotenv
numpy
